Figure out an optimal way to deal with precision in exchange rates and conversion

+++ src/test/kotlin/lu/mkremer/fundstransfer/AbstractIntegrationTest.kt
@@ -75,7 +75,7 @@ abstract class AbstractIntegrationTest {
     /**
      * Specifies a mocked set of exchange rates to be used by the system
      */
-    protected final fun mockExchangeRates(rates: Map<String, Double>) {
+    protected final fun mockExchangeRates(rates: Map<String, BigDecimal>) {
         Mockito.`when`(exchangeRateSynchronizer.fetch()).thenReturn(ExchangeRates(rates))
         updateExchangeRates()
     }
